Vaccination Policy
A vaccination policy was implemented. Vaccines are scheduled from a start date to a final day in a linear fashion. Distribution of types of vaccines gathered from a survey was considered, so that vaccine types (Pfizer, Astra Zeneca, etc.) are assigned according to that distribution. Type of vaccine determines efficacies for contagion and critical illness in agents.